+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Unreleased
 
+- Added pub key, priv key and mnemonic to Wallet struct.
 - Add `other: OtherFields` to `TransactionReceipt` [#2209[](https://github.com/gakonst/ethers-rs/pull/2209)
 - Add `Signature::recover_typed_data` [#2120](https://github.com/gakonst/ethers-rs/pull/2120)
 - Add `abi::encode_packed` [#2104](https://github.com/gakonst/ethers-rs/pull/2104)